+++ hyundai_kia_connect_api/ApiImplType1.py
@@ -8,13 +8,14 @@ from typing import Optional
 from time import sleep
 
 
-from .ApiImpl import ApiImpl, ScheduleChargingClimateRequestOptions
+from .ApiImpl import ApiImpl, ScheduleChargingClimateRequestOptions, ClimateRequestOptions
 from .Token import Token
 from .Vehicle import Vehicle
 
 from .utils import (
     get_child_value,
     parse_datetime,
+    get_index_into_hex_temp
 )
 
 from .const import (
@@ -701,6 +702,89 @@ class ApiImplType1(ApiImpl):
             url, json=payload, headers=self._get_control_headers(token, vehicle)
         ).json()
         _LOGGER.debug(f"{DOMAIN} - Schedule Charging and Climate Response: {response}")
+        _check_response_for_errors(response)
+        token.device_id = self._get_device_id(self._get_stamp())
+        return response["msgId"]
+
+    def start_climate(
+        self, token: Token, vehicle: Vehicle, options: ClimateRequestOptions
+    ) -> str:
+        url = self.SPA_API_URL + "vehicles/" + vehicle.id + "/control/temperature"
+
+        # Defaults are located here to be region specific
+
+        if options.set_temp is None:
+            options.set_temp = 21
+        if options.duration is None:
+            options.duration = 5
+        if options.defrost is None:
+            options.defrost = False
+        if options.climate is None:
+            options.climate = True
+        if options.heating is None:
+            options.heating = 0
+        if not vehicle.ccu_ccs2_protocol_support:
+            hex_set_temp = get_index_into_hex_temp(
+                self.temperature_range.index(options.set_temp)
+            )
+
+            payload = {
+                "action": "start",
+                "hvacType": 0,
+                "options": {
+                    "defrost": options.defrost,
+                    "heating1": int(options.heating),
+                },
+                "tempCode": hex_set_temp,
+                "unit": "C",
+            }
+            _LOGGER.debug(f"{DOMAIN} - Start Climate Action Request: {payload}")
+            response = requests.post(
+                url,
+                json=payload,
+                headers=self._get_authenticated_headers(
+                    token, vehicle.ccu_ccs2_protocol_support
+                ),
+            ).json()
+        _LOGGER.debug(f"{DOMAIN} - Start Climate Action Response: {response}")
+        _check_response_for_errors(response)
+        token.device_id = self._get_device_id(self._get_stamp())
+        return response["msgId"]
+
+    def stop_climate(self, token: Token, vehicle: Vehicle) -> str:
+        if not vehicle.ccu_ccs2_protocol_support:
+            url = self.SPA_API_URL + "vehicles/" + vehicle.id + "/control/temperature"
+            payload = {
+                "action": "stop",
+                "hvacType": 0,
+                "options": {
+                    "defrost": True,
+                    "heating1": 1,
+                },
+                "tempCode": "10H",
+                "unit": "C",
+            }
+            _LOGGER.debug(f"{DOMAIN} - Stop Climate Action Request: {payload}")
+            response = requests.post(
+                url,
+                json=payload,
+                headers=self._get_authenticated_headers(
+                    token, vehicle.ccu_ccs2_protocol_support
+                ),
+            ).json()
+        else:
+            url = self.SPA_API_URL_V2 + "vehicles/" + vehicle.id + "/ccs2/control/temperature"
+            payload = {
+                "command": "stop",
+                
+            }
+            _LOGGER.debug(f"{DOMAIN} - Stop Climate Action Request: {payload}")
+            response = requests.post(
+                url,
+                json=payload,
+                headers=self._get_control_headers(token, vehicle),
+            ).json()
+        _LOGGER.debug(f"{DOMAIN} - Stop Climate Action Response: {response}")
         _check_response_for_errors(response)
         token.device_id = self._get_device_id(self._get_stamp())
         return response["msgId"]
